requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

numpy==1.23.1
Pillow==9.2.0
sympy==1.11.1
